Delete non-existent Drupal 8 filter
In Drupal 8, `link` is a Twig function, not a filter. Furthermore,
although I haven't tested it, I don't think this code would work as
written, since the second argument to a Twig.js filter (`url`, in this
case) is actually an _array_ of the parameters passed into the filter at
calltime.